fix: middleware file upload

+++ internal/service/media.go
@@ -1,32 +1,31 @@
 package service
 
 import (
-	"errors"
+	"context"
 	"fmt"
 	"strings"
 
 	"go-rest-api/domain"
 	"go-rest-api/dto"
 	"go-rest-api/internal/config"
-	"io"
-	"mime/multipart"
 	"os"
 	"path/filepath"
-	"strconv"
 	"time"
 
 	"github.com/google/uuid"
 )
 
 type mediaService struct {
-	mediaRepo domain.MediaRepository
-	config    *config.Config
+	mediaRepo   domain.MediaRepository
+	bookService domain.BookService
+	config      *config.Config
 }
 
-func NewMediaService(mediaRepo domain.MediaRepository, config *config.Config) domain.MediaService {
+func NewMediaService(mediaRepo domain.MediaRepository, bookService domain.BookService, config *config.Config) domain.MediaService {
 	return &mediaService{
-		mediaRepo: mediaRepo,
-		config:    config,
+		mediaRepo:   mediaRepo,
+		bookService: bookService,
+		config:      config,
 	}
 }
 
@@ -64,55 +63,9 @@ func (s *mediaService) GetMediaByFileName(fileName string) (*dto.MediaResponse, 
 	return &response, nil
 }
 
-func (s *mediaService) UploadMedia(file *multipart.FileHeader) (*dto.MediaResponse, error) {
-	maxSize, err := strconv.ParseInt(s.config.File.MaxUploadSize, 10, 64)
-	if err != nil {
-		maxSize = 5
-	}
-	maxSize = maxSize * 1024 * 1024
-
-	if file.Size > maxSize {
-		return nil, errors.New("file size exceeds maximum upload size")
-	}
-
-	allowedExtensions := map[string]bool{
-		".jpg":  true,
-		".jpeg": true,
-		".png":  true,
-		".gif":  true,
-	}
-
-	ext := strings.ToLower(filepath.Ext(file.Filename))
-	if !allowedExtensions[ext] {
-		return nil, errors.New("unsupported file type")
-	}
-
-	uploadPath := s.config.File.UploadPath
-	if err := os.MkdirAll(uploadPath, os.ModePerm); err != nil {
-		return nil, fmt.Errorf("failed to create upload directory: %w", err)
-	}
-
-	newFileName := uuid.New().String() + ext
-	filePath := filepath.Join(uploadPath, newFileName)
-
-	src, err := file.Open()
-	if err != nil {
-		return nil, fmt.Errorf("failed to open uploaded file: %w", err)
-	}
-	defer src.Close()
-
-	dst, err := os.Create(filePath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create destination file: %w", err)
-	}
-	defer dst.Close()
-
-	if _, err = io.Copy(dst, src); err != nil {
-		return nil, fmt.Errorf("failed to copy file: %w", err)
-	}
-
+func (s *mediaService) UploadMedia(newFileName string, filePath string) (*dto.MediaResponse, error) {
 	media := &domain.Media{
-		Path:      s.config.File.LinkCover + "/" + newFileName,
+		Path:      newFileName,
 		CreatedAt: time.Now(),
 		UpdatedAt: time.Now(),
 	}
@@ -126,18 +79,34 @@ func (s *mediaService) UploadMedia(file *multipart.FileHeader) (*dto.MediaRespon
 	return &response, nil
 }
 
-func (s *mediaService) DeleteMedia(id uuid.UUID) error {
+func (s *mediaService) DeleteMedia(ctx context.Context, id uuid.UUID) error {
 	media, err := s.mediaRepo.FindByID(id)
 	if err != nil {
 		return err
 	}
 
-	filePath := filepath.Join(s.config.File.UploadPath, media.Path)
-	if err := os.Remove(filePath); err != nil && !os.IsNotExist(err) {
-		return fmt.Errorf("failed to delete file: %w", err)
+	books, err := s.bookService.GetBookByCoverID(ctx, media.ID)
+	if err != nil {
+		return err
 	}
 
-	return s.mediaRepo.Delete(id)
+	for _, book := range books {
+		err := s.bookService.DeleteBookCover(ctx, book.ID)
+		if err != nil {
+			return fmt.Errorf("failed to delete book cover for book ID %s: %w", book.ID, err)
+		}
+	}
+
+	errDelete := s.mediaRepo.Delete(id)
+	if errDelete == nil {
+		fileName := strings.TrimPrefix(media.Path, s.config.File.LinkCover+"/")
+		filePath := filepath.Join(s.config.File.UploadPath, fileName)
+		if err := os.Remove(filePath); err != nil && !os.IsNotExist(err) {
+			return fmt.Errorf("failed to delete file: %w", err)
+		}
+	}
+
+	return errDelete
 }
 
 func (s *mediaService) toMediaResponse(media *domain.Media) dto.MediaResponse {
